switch to 32-bit ints

+++ src/hanayo/file.rs
@@ -88,7 +88,7 @@ fn write(file: Value::Record, buf: Value::Str) -> Value {
     let file = file.as_mut();
     if let Some(field) = file.native_field.as_mut() {
         let file = field.downcast_mut::<File>().unwrap();
-        Value::Int(file.write_all(buf.as_ref().as_bytes()).is_ok() as i64)
+        Value::Int(file.write_all(buf.as_ref().as_bytes()).is_ok() as i32)
     } else {
         Value::Int(0)
     }
@@ -100,8 +100,8 @@ fn seek(file: Value::Record, pos: Value::Int) -> Value {
     let file = file.as_mut();
     if let Some(field) = file.native_field.as_mut() {
         let file = field.downcast_mut::<File>().unwrap();
-        if let Ok(result) = file.seek(SeekFrom::Current(pos)) {
-            Value::Int(result as i64)
+        if let Ok(result) = file.seek(SeekFrom::Current(pos as i64)) {
+            Value::Int(result as i32)
         } else {
             Value::Int(-1)
         }
@@ -116,7 +116,7 @@ fn seek_from_start(file: Value::Record, pos: Value::Int) -> Value {
     if let Some(field) = file.native_field.as_mut() {
         let file = field.downcast_mut::<File>().unwrap();
         if let Ok(result) = file.seek(SeekFrom::Start(pos as u64)) {
-            Value::Int(result as i64)
+            Value::Int(result as i32)
         } else {
             Value::Int(-1)
         }
@@ -130,8 +130,8 @@ fn seek_from_end(file: Value::Record, pos: Value::Int) -> Value {
     let file = file.as_mut();
     if let Some(field) = file.native_field.as_mut() {
         let file = field.downcast_mut::<File>().unwrap();
-        if let Ok(result) = file.seek(SeekFrom::End(pos)) {
-            Value::Int(result as i64)
+        if let Ok(result) = file.seek(SeekFrom::End(pos as i64)) {
+            Value::Int(result as i32)
         } else {
             Value::Int(-1)
         }
